RDKB-40723: Wan is starting before docsis is operational during wan mode
Reason for change:
1. Wan is starting before docsis is operational and
  reboot Issue can occur during software downgrade/upgrade.
  fix,
 Hence reboot check is done based on feature flag.
 Wan is starting after docsis is operational.

Test Procedure:
1. Test ipv6/ipv6 connectivity wan and lan.
2. Test Brigde mode and docsis mode, Ethwan mode.
3. Mac address leak in cpe table.

Risks: None
(cherry picked from commit 0e2ab601383f5700e34007f3dddb9d5097b36c99)
(cherry picked from commit 1bcf6bcca4313892b107a031eaed13151472dfa5)

+++ source/WanManager/wanmgr_policy_autowan_impl.c
@@ -724,7 +724,22 @@ static WcFmobPolicyState_t Transition_WaitingForInterface(WanMgr_Iface_Data_t *p
 
 static WcFmobPolicyState_t Transition_WanInterfaceTearDown(WanMgr_Policy_Controller_t* pWanController)
 {
+    DML_WAN_IFACE* pFixedInterface = NULL;
     WcFmobPolicyState_t retState = STATE_WAN_INTERFACE_TEARDOWN;
+
+    if((pWanController != NULL) && (pWanController->pWanActiveIfaceData != NULL))
+    {
+        pFixedInterface = &(pWanController->pWanActiveIfaceData->data);
+    }
+
+    if(pFixedInterface == NULL)
+    {
+        return retState;
+    }
+
+    // Reset Physical link status when state machine is teardown
+    pFixedInterface->Phy.Status = WAN_IFACE_PHY_STATUS_UNKNOWN;
+
     wanmgr_setwanstop();
     system("killall dibbler-client");
     system("killall udhcpc");
@@ -851,6 +866,10 @@ static WcFmobPolicyState_t Transition_WanInterfaceActive(WanMgr_AutoWan_SMInfo_t
 
     CcspTraceInfo(("%s %d - State changed to STATE_WAN_INTERFACE_ACTIVE if_name %s\n", __FUNCTION__, __LINE__,pFixedInterface->Wan.Name));
     CcspTraceInfo(("%s %d - LastKnownMode %d  Active index %d\n", __FUNCTION__, __LINE__,lastKnownMode,pWanController->activeInterfaceIdx));
+#ifndef ENABLE_WANMODECHANGE_NOREBOOT
+    // Do reboot during wan mode change if NOREBOOT feature is not enabled.
+    pFixedInterface->Wan.RebootOnConfiguration = TRUE;
+#endif
     switch (lastKnownMode)
     {
         case WAN_MODE_PRIMARY:
